delete handler refactoring
