+++ Sources/VaporRestKit/Deprecated/Controllers/RelatedResourceControllers/ParentChild/ChildResourceControllers.swift
@@ -108,7 +108,7 @@ struct DeleteChildrenResourceController<Model, RelatedModel, Output, EagerLoadin
 
 
     let relatedResourceMiddleware: ControllerMiddleware<Model, RelatedModel>
-    let deleter: DeleteHandler<Model>
+    let deleter: Deleter<Model>
     let relationNamePath: String?
     let childrenKeyPath: ChildrenKeyPath<RelatedModel, Model>
