+++ Tests/VaporRestKitTests/Controllers/GalaxyControllerV2.swift
@@ -41,7 +41,7 @@ struct GalaxyControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Galaxy.Output> {
             try ResourceController<Galaxy.Output>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier)
         }
 
@@ -93,7 +93,7 @@ struct GalaxyControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Galaxy.Output> {
             try RelatedResourceController<Galaxy.Output>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \Galaxy.$stars)
         }
@@ -148,7 +148,7 @@ struct GalaxyControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Galaxy.ExtendedOutput<Star.Output>> {
             try RelatedResourceController<Galaxy.ExtendedOutput<Star.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \Galaxy.$stars)
         }
