+++ Tests/VaporRestKitTests/Controllers/StarControllerV2.swift
@@ -44,7 +44,7 @@ struct StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.Output> {
             try ResourceController<Star.Output>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier)
         }
 
@@ -93,7 +93,7 @@ struct StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try ResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier)
         }
 
@@ -142,7 +142,7 @@ struct StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try ResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier)
         }
 
@@ -191,7 +191,7 @@ struct StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try ResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier)
         }
 
@@ -244,7 +244,7 @@ struct StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.Output> {
             try RelatedResourceController<Star.Output>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \StarTag.$stars)
         }
@@ -299,7 +299,7 @@ struct StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try RelatedResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \StarTag.$stars)
         }
@@ -354,7 +354,7 @@ struct StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try RelatedResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \StarTag.$stars)
         }
@@ -409,7 +409,7 @@ struct StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try RelatedResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \StarTag.$stars)
         }
@@ -483,7 +483,7 @@ extension StarControllersV2  {
             try RelatedResourceController<Star.Output>().delete(
                 
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \Star.$planets)
         }
@@ -540,7 +540,7 @@ extension StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try RelatedResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \Galaxy.$stars)
         }
@@ -595,7 +595,7 @@ extension StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try RelatedResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \Galaxy.$stars)
         }
@@ -650,7 +650,7 @@ extension StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>> {
             try RelatedResourceController<Star.ExtendedOutput<Galaxy.Output, StarTag.Output>>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \Galaxy.$stars)
         }
@@ -706,7 +706,7 @@ extension StarControllersV2 {
         func delete(req: Request) throws -> EventLoopFuture<Star.Output> {
             try RelatedResourceController<Star.Output>().delete(
                 req: req,
-                using: .defaultDeleter,
+                using: .defaultDeleter(),
                 queryModifier: queryModifier,
                 relationKeyPath: \Galaxy.$stars)
         }
